rename TestRetries::Strategy to TestRetries::Driver
