+++ lib/datadog/ci/test_retries/driver/no_retry.rb
@@ -5,7 +5,7 @@ require_relative "base"
 module Datadog
   module CI
     module TestRetries
-      module Strategy
+      module Driver
         class NoRetry < Base
           def record_retry(test_span)
           end
