+++ sig/datadog/ci/test_retries/driver/no_retry.rbs
@@ -1,13 +1,9 @@
 module Datadog
   module CI
     module TestRetries
-      module Strategy
-        class Base
-          def should_retry?: () -> bool
-
+      module Driver
+        class NoRetry < Base
           def record_retry: (Datadog::CI::Test test_span) -> void
-
-          def record_duration: (Float duration) -> void
         end
       end
     end
