+++ sig/datadog/ci/test_retries/driver/retry_failed.rbs
@@ -1,7 +1,7 @@
 module Datadog
   module CI
     module TestRetries
-      module Strategy
+      module Driver
         class RetryFailed < Base
           attr_reader max_attempts: Integer
