+++ spec/datadog/ci/test_retries/driver/retry_failed_spec.rb
@@ -1,11 +1,11 @@
-require_relative "../../../../../lib/datadog/ci/test_retries/strategy/retry_failed"
+require_relative "../../../../../lib/datadog/ci/test_retries/driver/retry_failed"
 
-RSpec.describe Datadog::CI::TestRetries::Strategy::RetryFailed do
+RSpec.describe Datadog::CI::TestRetries::Driver::RetryFailed do
   let(:max_attempts) { 3 }
-  subject(:strategy) { described_class.new(max_attempts: max_attempts) }
+  subject(:driver) { described_class.new(max_attempts: max_attempts) }
 
   describe "#should_retry?" do
-    subject { strategy.should_retry? }
+    subject { driver.should_retry? }
 
     context "when the test has not passed yet" do
       let(:test_span) { double(:test_span, set_tag: true, passed?: false) }
@@ -13,7 +13,7 @@ RSpec.describe Datadog::CI::TestRetries::Strategy::RetryFailed do
       it { is_expected.to be true }
 
       context "when the max attempts have been reached" do
-        before { max_attempts.times { strategy.record_retry(test_span) } }
+        before { max_attempts.times { driver.record_retry(test_span) } }
 
         it { is_expected.to be false }
       end
@@ -22,7 +22,7 @@ RSpec.describe Datadog::CI::TestRetries::Strategy::RetryFailed do
     context "when the test has passed" do
       let(:test_span) { double(:test_span, set_tag: true, passed?: true) }
 
-      before { strategy.record_retry(test_span) }
+      before { driver.record_retry(test_span) }
 
       it { is_expected.to be false }
     end
