+++ spec/datadog/ci/test_retries/driver/retry_new_spec.rb
@@ -1,6 +1,6 @@
-require_relative "../../../../../lib/datadog/ci/test_retries/strategy/retry_new"
+require_relative "../../../../../lib/datadog/ci/test_retries/driver/retry_new"
 
-RSpec.describe Datadog::CI::TestRetries::Strategy::RetryNew do
+RSpec.describe Datadog::CI::TestRetries::Driver::RetryNew do
   let(:max_attempts) { 10 }
   let(:duration_thresholds) {
     Datadog::CI::Remote::SlowTestRetries.new({
@@ -12,29 +12,29 @@ RSpec.describe Datadog::CI::TestRetries::Strategy::RetryNew do
   }
   let(:test_span) { double(:test_span, set_tag: true) }
 
-  subject(:strategy) { described_class.new(test_span, duration_thresholds: duration_thresholds) }
+  subject(:driver) { described_class.new(test_span, duration_thresholds: duration_thresholds) }
 
   describe "#should_retry?" do
-    subject { strategy.should_retry? }
+    subject { driver.should_retry? }
 
     context "when max attempts haven't been reached yet" do
       it { is_expected.to be true }
     end
 
     context "when the max attempts have been reached" do
-      before { max_attempts.times { strategy.record_retry(test_span) } }
+      before { max_attempts.times { driver.record_retry(test_span) } }
 
       it { is_expected.to be false }
     end
   end
 
   describe "#record_duration" do
-    subject { strategy.record_duration(duration) }
+    subject { driver.record_duration(duration) }
 
     let(:duration) { 5 }
 
     it "updates the max attempts based on the duration" do
-      expect { subject }.to change { strategy.instance_variable_get(:@max_attempts) }.from(10).to(5)
+      expect { subject }.to change { driver.instance_variable_get(:@max_attempts) }.from(10).to(5)
     end
   end
 end
